src/requirements.txt
alabaster==0.7.12
aniso8601==8.1.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docutils==0.16
dotmap==1.3.23
ephem==3.7.7.1
Flask==1.1.2
Flask-API==2.0
Flask-RESTful==0.3.8
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
Marshmallow-Peewee==3.0.0
packaging==20.9
peewee==3.14.1
pluggy==0.13.1
py==1.10.0
Pygments==2.7.4
pylunar==0.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
Sphinx==3.4.3
sphinx-py3doc-enhanced-theme==2.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
